fix(testnet): metadata api
