+++ runtime/testnet/Cargo.toml
@@ -107,6 +107,7 @@ pallet-ismp-runtime-api.workspace = true
 [dev-dependencies]
 enumflags2.workspace = true
 env_logger.workspace = true
+frame-metadata.workspace = true
 hex.workspace = true
 
 [features]
